+++ CHANGELOG.rst
@@ -7,9 +7,16 @@ Changelog
 
 Added
 ~~~~~
+- When using phaser, a LLG/TFZ > 120/8 was added as an additional criteria for early termination
+- Added a process all flag to amore/phaser rotation functions so that they skip the early termination funciton
 
 Changed
 ~~~~~~~
+- Default sequence identity changed from 30 percent to 70
+- Fixed a test case that was broken
+- Fixed several bugs when running jobs as multiple chunks
+- Changed process to skip successive chunks if a solution is found
+- Fixed path in pyvapi
 
 0.1.16
 ------